[refactor] Split agentfile.go into specs/install/block/status (#146)
Pure file-move (Fowler Split Module): no symbol renamed, no signature changed,
no behaviour change. agentfile.go becomes the types-and-constants anchor;
install/uninstall machinery moves to install.go, spec catalogs to specs.go,
status checks to status.go, block helpers to block.go.

Test files split symmetrically: install_test.go, specs_test.go, status_test.go,
block_test.go, content_test.go, with agentfile_test.go reduced to shared
test-helper constants.

Closes #146
